correções

+++ app/Http/Controllers/Painel/CustomerController.php
@@ -176,8 +176,10 @@ class CustomerController extends Controller
      */
     public function destroy($id)
     {
-        $customer = Customer::find($id);
-        $customer->delete();
+        if(!empty($id)){
+            $customer = Customer::find($id);
+            $customer->delete();
+        }
         return redirect()->route('customers.index');
     }
 }
